refactor(app): move screen flows into app and remove handler module
This commit relocates screen-handling and loading logic under the app
boundary and inlines the former run_app loop inside AppActor. The handler
module is removed so application orchestration lives entirely within the app
crate, with per-screen flows dispatched from the actor's input path.

This commit is part of the architecture's refactoring phase 12.
