+++ src/app/flows/details_actions.rs
