+++ src/app/flows/mod.rs
@@ -1,0 +1,5 @@
+pub(crate) mod bookmarked;
+pub(crate) mod details_actions;
+pub(crate) mod edit_config;
+pub(crate) mod latest;
+pub(crate) mod mail_list;
